+++ src/plots/plots_batch.R
@@ -84,7 +84,7 @@ makeOneMultiPlotInBatch <- function(mpConf, exptDb) {
   # Making the plots
   plots <- prepareMultiPlots(
     plotter=plotTypesFlat[[mpConf$plotType]],
-    inputsForAllPlots=inputsForAllPlots, db=exptDb
+    inputsForAllPlots=inputsForAllPlots, interactive=FALSE, db=exptDb
   )
 
   fileType <- bmConf$fileType
@@ -93,7 +93,7 @@ makeOneMultiPlotInBatch <- function(mpConf, exptDb) {
       '  > multiPlot "%s": Saving plot %d of %d...',
       mpConf$displayName, iPlt, length(plots)
     )
-    plot <- addTitleToPlot(plots[[iPlt]]$obplot, plots[[iPlt]]$title)
+    plot <- plots[[iPlt]]$obplot %>% addTitleToPlot(plots[[iPlt]]$title)
     fName <- file.path(dirPath, sprintf("plot_%s.%s",iPlt,bmConf$fileType))
     ggsave(
       filename=fName, plot=plot, device=bmConf$fileType,