+++ src/plots/plots_maps.R
@@ -25,7 +25,6 @@ plotTitle.plotMap <- function(p, plotRequest, plotData) {
   return(title)
 }
 
-
 doPlot.plotMap <- function(p, plotRequest, plotData) {
   x1 <- min(plotData$longitude)-2
   x2 <- max(plotData$longitude)+2
@@ -37,6 +36,47 @@ doPlot.plotMap <- function(p, plotRequest, plotData) {
               colour="gray50") +
     coord_map("stereographic", xlim=c(x1, x2), ylim=c(y1, y2)) +
     labs(x="Longitude", y="Latitude")
+}
+
+doPlotly.plotMap <- function(p, plotRequest, plotData) {
+  myPlotly <- plot_geo(
+    data=plotData, lat=~latitude, lon =~longitude
+  ) %>%
+    layout(
+      margin = list(
+        t=100, # To leave space for the title
+        b=10, # Looks better when figure is exported
+        l=175, r=175 # To prevent legend from ending up too far away from plot
+      ),
+      showlegend = TRUE,
+      legend = list(
+        orientation="v", yanchor="center", y=0.5, xanchor="left", x=1.01
+      ),
+      geo = list(
+        # See <https://plot.ly/r/reference/#layout-geo>
+        resolution = 50,
+        showland = TRUE,
+        showlakes = TRUE,
+        showcountries=TRUE,
+        landcolor = toRGB("grey98"),
+        countrycolor = toRGB("grey50"),
+        lakecolor = toRGB("white"),
+        projection = list(type="stereographic"),
+        coastlinewidth = 0.5,
+        countrywidth = 0.5,
+        lataxis = list(
+          range = range(plotData$latitude) + c(-1, 1),
+          showgrid = TRUE,
+          dtick = 10
+        ),
+        lonaxis = list(
+          range = range(plotData$longitude) + c(-2, 2),
+          showgrid = TRUE,
+          dtick = 15
+        )
+      )
+    )
+  return(myPlotly)
 }
 
 doMap.plotMap <- function(p, plotRequest, plotData) {
@@ -131,10 +171,9 @@ doMap.mapThreshold <- function(p, plotRequest, plotData) {
   }else{
     plotData$radius=5
   }
-  cm <- colormapContinuous(min(plotData$plotValues),
-                           max(plotData$plotValues))
-  dataPal <- colorNumeric(palette=cm$palette, domain=cm$domain, reverse=cm$reverse)
-  legendPal <- colorNumeric(palette=cm$palette, domain=cm$domain, reverse=!cm$reverse)
+  cm <- getSuitableColorScale(plotData)
+  dataPal <- colorNumeric(palette=cm$palette, domain=cm$domain)
+  legendPal <- colorNumeric(palette=rev(cm$palette), domain=cm$domain)
   clusterOptions <- markerClusterOptions(disableClusteringAtZoom=zoomLevel)
   obMap <- leaflet(plotData) %>%
     addProviderTiles("Esri.WorldStreetMap",
@@ -175,8 +214,12 @@ doMap.mapUsage <- function(p, plotRequest, plotData) {
   pal <- colorFactor(c("green", "blue", "black", "grey", "magenta3", "red"),
                      domain=c("Active", "Active(2)",
                               "Blacklisted", "NA", "Passive", "Rejected"))
-  plotData$popup <- paste("Station: ", plotData$statLabel, "<br>Anflag: ",
-                          plotData$anflag, "<br>Status:", plotData$status)
+  plotData$popup <- paste(
+                      "Station: ", plotData$statLabel,
+                      "<br>Level: ", plotData$level,
+                      "<br>Anflag: ", plotData$anflag,
+                      "<br>Status:", plotData$status
+                    )
   clusterOptions <- markerClusterOptions(disableClusteringAtZoom=zoomLevel)
   obMap <- leaflet(data=plotData[rev(order(status)),]) %>%
     addProviderTiles("Esri.WorldStreetMap",
@@ -221,12 +264,56 @@ doPlot.mapUsage <- function(p, plotRequest, plotData) {
                                  "Blacklisted"="black", "NA"="grey"))
 }
 
+doPlotly.mapUsage <- function(p, plotRequest, plotData) {
+  status <- rep("NA", nrow(plotData))
+  status <- ifelse(plotData$anflag == 0, "Rejected", status)
+  status <- ifelse(plotData$active  > 0, "Active", status)
+  status <- ifelse(plotData$rejected > 0, "Rejected", status)
+  status <- ifelse(plotData$passive > 0, "Passive", status)
+  status <- ifelse(plotData$blacklisted > 0, "Blacklisted", status)
+  status <- ifelse(plotData$anflag  > 0, "Active(2)", status)
+  status <- ifelse(plotData$anflag == 4, "Rejected", status)
+  status <- ifelse(plotData$anflag == 8, "Blacklisted", status)
+  plotData$status <- status
+
+  colors <- c(
+    "Active"="green", "Active(2)"="blue",
+    "Rejected"="red", "Passive"="magenta3",
+    "Blacklisted"="black", "NA"="grey"
+  )
+
+  myPlotly <- NextMethod()
+  myPlotly <- myPlotly %>%
+    add_markers(
+      data=plotData[rev(order(status)),],
+      text=~paste(
+        paste("Station:", statLabel),
+        paste("Level:", level),
+        sprintf("Coords: (%.3f\u00B0, %.3f\u00B0)", longitude, latitude),
+        paste("Anflag:", anflag),
+        paste("Status:", status),
+        sep="<br />"
+      ),
+      marker = list(
+        line = list(
+          color = 'black',
+          width = 1,
+          opacity=0.5
+        )
+      ),
+      symbol=~status, color=~status, colors=colors, size=2,
+      hoverinfo="text"
+    )
+
+  return(myPlotly)
+}
+
 registerPlotType(
     "Maps",
     plotCreate(c("mapUsage", "plotMap"),
                "Observation Usage", "single",
                paste("SELECT",
-                     "latitude, longitude, statid,",
+                     "latitude, longitude, level, statid,",
                      "active, rejected, passive, blacklisted, anflag, obsvalue",
                      "FROM usage WHERE %s"),
                list("obnumber", "obname"))
@@ -239,13 +326,44 @@ plotBuildQuery.mapThreshold <- function(p, plotRequest) {
 doPlot.mapThreshold <- function(p, plotRequest, plotData) {
   minval <- min(plotData$plotValues)
   maxval <- max(plotData$plotValues)
-  cm <- colormapContinuous(minval, maxval)
+  cm <- getSuitableColorScale(plotData)
   NextMethod() +
     geom_point(data=plotData,
                aes(x=longitude, y=latitude, fill=plotValues),
                size=3, shape=21, colour="gray50", alpha=.5, stroke=0.) +
-    scale_fill_distiller(p$dataColumn, type=cm$type, palette=cm$palette,
+    scale_fill_distiller(p$dataColumn, palette=cm$name,
                          direction=cm$direction, limits=cm$domain)
+}
+
+doPlotly.mapThreshold <- function(p, plotRequest, plotData) {
+  cm <- getSuitableColorScale(plotData)
+  myPlotly <- NextMethod()
+  myPlotly <- myPlotly %>%
+    add_markers(
+      text=~paste(
+        paste("Station:", statLabel),
+        paste("Level:", level),
+        sprintf("Coords: (%.3f\u00B0, %.3f\u00B0)", longitude, latitude),
+        paste0(p$dataColumn, ": ", signif(plotValues, digits=5)),
+        sep="<br />"
+      ),
+      size=2, color=~plotValues, colors=cm$palette,
+      marker = list(
+        line = list(
+          color = 'black',
+          width = 1,
+          opacity=0.5
+        )
+      ),
+      hoverinfo="text"
+    ) %>%
+    colorbar(
+      limits=cm$domain,
+      title=p$dataColumn,
+      yanchor="center", y=0.5,
+      xanchor="left", x=1.0
+    )
+  return(myPlotly)
 }
 
 registerPlotType(
@@ -323,17 +441,17 @@ registerPlotType(
 registerPlotCategory("AverageMaps")
 
 mapThresholdWithRangeAggregateAndApplyFunction <-
-  function(plotter, plotData, FUN='mean', 
+  function(plotter, plotData, FUN='mean',
     aggregateBy=c("statid", "latitude", "longitude", "level")
   ) {
-  # Grouping data by the colnames specified in aggregateBy, 
+  # Grouping data by the colnames specified in aggregateBy,
   # then applying function FUN within each group
   if(isTRUE(nrow(plotData)>0)) {
     aggregateByList = plotData[, aggregateBy]
-    columnsNotToBeAggreg <- which(colnames(plotData) %in% 
+    columnsNotToBeAggreg <- which(colnames(plotData) %in%
                                 c("DTG", aggregateBy)
                               )
-    plotData <- aggregate(plotData[, -columnsNotToBeAggreg], 
+    plotData <- aggregate(plotData[, -columnsNotToBeAggreg],
                   by=aggregateByList,
                   FUN=FUN,
                   na.rm=TRUE
@@ -346,7 +464,7 @@ mapThresholdWithRangeAggregateAndApplyFunction <-
 postProcessQueriedPlotData.mapThresholdWithRangeAvgs <-
   function(plotter, plotData) {
     mapThresholdWithRangeAggregateAndApplyFunction(
-      plotter, plotData, FUN="mean", 
+      plotter, plotData, FUN="mean",
       aggregateBy=c("statid", "latitude", "longitude", "level")
   )
 }